API Reference

Ray Types

RayTracing.AbstractRay — Type

abstract type AbstractRay{ND, T <: AbstractFloat}

Represents a ray in ND dimensions that uses the type T for numeric calculations.

All rays have a position and a direction, and each concrete type may carry additional information.

RayTracing.BasicRay — Type

struct BasicRay{ND, T} <: AbstractRay{ND, T}

A basic implementation of a ray in ND dimensions with arithmetic type T.

Only stores a current position and a direction.

RayTracing.BasicRay — Method

BasicRay{ND, T}(pos, dir) where {ND, T}
BasicRay(pos, dir)

Construct a BasicRay{ND, T} with position pos and direction dir. With the second signature, ND and T are determined automatically from the arguments.

The direction is normalized before costruction.

Both arguments should be vectors of length ND and element type T. Both are copied so that later changes to the objects passed as arguments don't affect the constructed ray.

RayTracing.StepRecorder — Type

struct StepRecorder{ND, T, R <: AbstractRay{ND, T}} <: AbstractRay{ND, T}

Wrapper around a ray type, R, that records the trajectory of the ray.

ND is the number of dimensions and T is the type used for numeric calculations.

The trajectory is stored in the field steps :: Vector{Point{ND, T}}. A new point is added to the trajectory after every call to advance!.

RayTracing.StepRecorder — Method

StepRecorder{ND, T, R}(ray :: R) where {ND, T, R}
StepRecorder(ray)

Construct a StepRecorder{ND, T, R} that wraps ray. With the second signature, ND, T and R are determined automatically from the argument.

Ray Interface

Base.eltype — Method

Base.eltype(::Type{<:AbstractRay})
Base.eltype(::AbstractRay)

Return the numeric type used by a ray object or a ray type.

Base.ndims — Method

Base.ndims(::Type{<:AbstractRay})
Base.ndims(::AbstractRay)

Return the number of dimensions of a ray object or a ray type.

RayTracing.advance! — Method

advance!(ray, distance)

Advances the ray by the given distance in its current direction.

RayTracing.direction — Function

direction(ray)

Return the current direction of the ray, possibly as a mutable reference.

RayTracing.position — Function

Base.position(ray)
RayTracing.position(ray)

Return the current position of the ray, possibly as a mutable reference.

RayTracing.reflect! — Method

reflect!(ray, normal)

Invert the direction of the ray with respect to the given normal vector.

normal should be a unit vector.

Scene Types

RayTracing.AbstractScene — Type

abstract type AbstractScene

A set of surfaces that rays can bounce on.

RayTracing.BallPit — Type

struct BallPit{ND, T} <: AbstractScene

A set of same-size spheres in ND dimensions with arithmetic type T.

The arguments for the constructor are: - r::T, the radius of the spheres. - centers::Vector{Point{ND, T}}, the centers of the spheres.

Scene Interface

RayTracing.trace! — Function

trace!(ray, scene)

Trace a ray on a scene.

The ray is evolved by advance!ing to each next intersection with an algorithm defined by each scene type.

RayTracing.trace! — Method

trace!(ray, scene::BallPit; max_steps=1000)

Trace ray on the given BallPit by advance!ing to each next intersection and reflect!ing on the corresponding sphere, until there are no more intersections or max_steps are reached.

It is assumed that the initial position of ray is not inside any sphere.