API Reference

Ray Types

RayTracing.AbstractRay — Type

abstract type AbstractRay{ND, T <: Real}

Represents a ray in ND dimensions that uses the type T for numeric calculations.

All rays have a position and a direction, and each concrete type may carry additional information.

RayTracing.BasicRay — Type

BasicRay{ND, T} <: AbstractRay{ND, T}

A basic implementation of a ray in ND dimensions with arithmetic type T.

Only stores a current position and a direction.

RayTracing.BasicRay — Method

BasicRay{ND, T}(pos, dir) where {ND, T}
BasicRay(pos, dir)

Construct a BasicRay{ND, T} with position pos and direction dir.

With the second signature, ND and T are determined automatically from the arguments.

The direction is normalized before costruction.

Both arguments should be vectors of length ND and element type T. Both are copied so that later changes to the objects passed as arguments don't affect the constructed ray.

Ray Interface

Base.eltype — Method

Base.eltype(::Type{<:AbstractRay})
Base.eltype(::AbstractRay)

Return the numeric type used by a ray object or a ray type.

Base.ndims — Method

Base.ndims(::Type{<:AbstractRay})
Base.ndims(::AbstractRay)

Return the number of dimensions of a ray object or a ray type.

RayTracing.advance! — Method

advance!(ray, distance)

Advances the ray by the given distance in its current direction.

RayTracing.direction — Function

direction(ray)

Return the current direction of the ray, as a mutable reference.

RayTracing.position — Function

Base.position(ray)
RayTracing.position(ray)

Return the current position of the ray, as a mutable reference.

Scene Types

RayTracing.AbstractScene — Type

abstract type AbstractScene

A set of objects that rays can bounce on.

Includes information about the geometry of the objects and also about surface properties.

See also: SceneGeometry.

RayTracing.SceneGeometry — Type

abstract type SceneGeometry

A geometric description of a set of objects.

Similar to AbstractScene, but only contains geometric information. Scenes can be created by adding surface information to existing scene geometries.

RayTracing.BallPit — Type

BallPit{ND, T} <: SceneGeometry

A set of same-size spheres in ND dimensions with arithmetic type T.

The arguments for the constructor are:

r::T, the radius of the spheres.

centers::Vector{Point{ND, T}}, the centers of the spheres.

Scene Interface